% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bichibarsq.R
\name{Bichibarsquare}
\alias{Bichibarsquare}
\alias{dbichibarsq}
\alias{pbichibarsq}
\alias{rbichibarsq}
\title{The Bivariate Chi-bar-squared Distribution}
\usage{
dbichibarsq(x, v)

pbichibarsq(x, v)

rbichibarsq(n, v)
}
\arguments{
\item{x}{two-entry vectors or two-column matrices.}

\item{v}{vector of mixing weights (conic intrinsic volumes).}

\item{n}{number of observations.}
}
\value{
The output of \code{dbichibarsq(x, v)} is either a number, if \code{x}
        is a vector of length \code{2}, or a vector of length \code{dim(x)[1]}
        such that the entries are the values of the density of the
        bivariate chi-bar-squared distribution with weights \code{v} in the
        corresponding entry (row) of \code{x}.

The output of \code{pbichibarsq(x, v)} is either a number, if \code{x}
        is a vector of length \code{2}, or a vector of length \code{dim(x)[1]}
        such that the entries are the values of the cdf of the
        bivariate chi-bar-squared distribution with weights \code{v} in the
        corresponding entry (row) of \code{x}.

The output of \code{rbichibarsq(n, v)} is either a number, if \code{n==1},
        or a two-column matrix with \code{n} rows such that the rows are
        iid samples from the bivariate chi-bar-squared distribution with weights \code{v}.
}
\description{
Density, distribution function, and random generation for the bivariate
chi-bar-squared distribution with mixing weights (conic intrinsic volumes) \code{v}.
}
\section{Functions}{
\itemize{
\item \code{dbichibarsq}: Evaluates the density of the bivariate chi-bar-squared distribution.

\item \code{pbichibarsq}: Evaluates the cdf of the bivariate chi-bar-squared distribution.

\item \code{rbichibarsq}: Samples from the bivariate chi-bar-squared distribution.
}}

\section{See also}{

\code{\link[conivol]{dchibarsq}}, \code{\link[conivol]{pchibarsq}}, \code{\link[conivol]{rchibarsq}},
\code{\link[stats]{dchisq}}, \code{\link[stats]{pchisq}}, \code{\link[stats]{qchisq}}, \code{\link[stats]{rchisq}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
dbichibarsq(c(2,3),c(0.1,0.6,0.3))
dbichibarsq(matrix(c(2,3,4,5),2,2),c(0.1,0.6,0.3))

pbichibarsq(c(2,3),c(0.1,0.6,0.3))
pbichibarsq(matrix(c(2,3,4,5),2,2),c(0.1,0.6,0.3))

rbichibarsq(10,c(0.1,0.6,0.3))

}
