% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chibarsq.R
\name{Chibarsquare}
\alias{Chibarsquare}
\alias{dchibarsq}
\alias{pchibarsq}
\alias{rchibarsq}
\title{The Chi-bar-squared Distribution}
\usage{
dchibarsq(x, v)

pchibarsq(q, v)

rchibarsq(n, v)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{v}{vector of mixing weights (conic intrinsic volumes).}

\item{n}{number of observations.}
}
\value{
The output of \code{dchibarsq(x, v)} is a vector of length \code{length(x)}
        such that the entries are the values of the density of the
        chi-bar-squared distribution with weights \code{v} in the
        corresponding entry of \code{x}.

The output of \code{pchibarsq(q, v)} is a vector of length \code{length(q)}
        such that the entries are the values of the cdf of the
        chi-bar-squared distribution with weights \code{v} in the
        corresponding entry of \code{q}.

The output of \code{rchibarsq(n, v)} is a vector of length \code{n}
        such that the entries are iid samples from the
        chi-bar-squared distribution with weights \code{v}.
}
\description{
Density, distribution function, and random generation for the chi-bar-squared
distribution with mixing weights (conic intrinsic volumes) \code{v}.
}
\section{Functions}{
\itemize{
\item \code{dchibarsq}: Evaluates the density of the chi-bar-squared distribution.

\item \code{pchibarsq}: Evaluates the cdf of the chi-bar-squared distribution.

\item \code{rchibarsq}: Samples from the chi-bar-squared distribution.
}}

\section{See also}{

\code{\link[conivol]{dbichibarsq}}, \code{\link[conivol]{pbichibarsq}}, \code{\link[conivol]{rbichibarsq}},
\code{\link[stats]{dchisq}}, \code{\link[stats]{pchisq}}, \code{\link[stats]{qchisq}}, \code{\link[stats]{rchisq}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
dchibarsq(3,c(0.1,0.6,0.3))
dchibarsq(c(0.2,3,4.5),c(0.1,0.6,0.3))

pchibarsq(3,c(0.1,0.6,0.3))
pchibarsq(c(0.2,3,4.5),c(0.1,0.6,0.3))

rchibarsq(10,c(0.1,0.6,0.3))

}
