% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_cones.R
\name{circ_ivol}
\alias{circ_ivol}
\title{The conic intrinsic volumes of (products of) circular cones.}
\usage{
circ_ivol(d, alpha, product = FALSE)
}
\arguments{
\item{d}{vector of dimensions; must be same length as \code{alpha}.}

\item{alpha}{vector of angles; must be same length as \code{d}.}

\item{product}{logical; if \code{TRUE}, intrinsic volumes of product cone are returned.}
}
\value{
If \code{length(d)==1} or \code{(length(d)>1 & product==TRUE)}
        then a single vectors will be returned.
        If \code{(length(d)>1 & product==FALSE)} then a list of
        vectors will be returned.
}
\description{
\code{circ_ivol} computes the conic intrinsic volumes of circular cones,
whose dimensions and angles are given in the vectors \code{d} and
\code{alpha} (vectors must be of same lengths); if the length of the vectors
is one, a single vector is returned; if the length of the vectors is greater
than one and \code{prcoduct==FALSE}, a list of vectors is returned;
if the length of the vectors is greater than one and \code{product==TRUE},
a single vector with the intrinsic volumes of the product cone is returned.
}
\section{See also}{

\code{\link[conivol]{circ_rbichibarsq}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
circ_ivol(5, pi/4)
circ_ivol(c(5,5), c(pi/4,pi/8))
circ_ivol(c(5,5), c(pi/4,pi/8), product = TRUE)

}
