% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_with_ivols.R
\name{comp_ivols_product}
\alias{comp_ivols_product}
\title{Compute intrinsic volumes of product cone from intrinsic volumes of components}
\usage{
comp_ivols_product(V)
}
\arguments{
\item{V}{list of vectors of intrinsic volumes}
}
\value{
The output of \code{comp_ivols_product(V)} is the vector that is
        obtained from convolving the components of \code{V}.
}
\description{
\code{comp_ivols_product} computes the intrinsic volumes of a product cone
from the intrinsic volumes of its components. That is, it returns the
convolution of the vectors of intrinsic volumes given in \code{V}.
}
\section{See also}{

Package: \code{\link[conivol]{conivol}}
}

\examples{
comp_ivols_product(list( c(0.5,0.5), c(0.1,0.4,0.5) ))
comp_ivols_product(list( c(0.5,0.5), c(0.5,0.5), c(0.5, 0.5) ))

}
