% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weights.R
\name{comp_loglike}
\alias{comp_loglike}
\title{Evaluate the log-likelihood of the estimated intrinsic volumes}
\usage{
comp_loglike(v, data, mode = 0)
}
\arguments{
\item{v}{vector of mixing weights (conic intrinsic volumes).}

\item{data}{output of \code{prepare_data_em(d, m_samp)}.}

\item{mode}{specifies whether the first and last values should be taken into account:
\describe{
  \item{\code{init_mode==0}:}{take all into account}
  \item{\code{init_mode==1}:}{leave out the estimate of the dth intrinsic volume}
  \item{\code{init_mode==2}:}{leave out the estimate of the 0th intrinsic volume}
  \item{\code{init_mode==3}:}{leave out both estimates of the 0th and dth intrinsic volume}
}}
}
\value{
The output of \code{comp_loglike} is the value of the normalized
        log-likelihood of the mixing weights \code{v} with respect to the
        sample data given in \code{data}
}
\description{
\code{comp_loglike} evaluates the (normalized) log-likelihood of a vector
with respect to given data, the output of \code{prepare_data_em}.
}
\section{See also}{

\code{\link[conivol]{prepare_data_em}}, \code{\link[conivol]{estimate_statdim_var}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
D <- c(5,5)
alpha <- c(pi/3,pi/4)
d <- sum(D)
N <- 10^5
v_exact <- circ_ivols( D, alpha, product=TRUE )

# collect sample data
m_samp <- circ_rbichibarsq(N,D,alpha)
data <- prepare_data_em(d, m_samp)
est <- estimate_statdim_var(d, m_samp)
v1 <- init_v( d )
v2 <- init_v( d, 1, delta=est$delta, var=est$var )
v3 <- init_v( d, 2, delta=est$delta )
v4 <- init_v( d, 3, var=est$var )
v5 <- init_v( d, 4, delta=est$delta, var=est$var )

# evaluate log-likelihood function
comp_loglike(v_exact, data)
comp_loglike(v1, data)
comp_loglike(v2, data)
comp_loglike(v3, data)
comp_loglike(v4, data)
comp_loglike(v5, data)

}
