% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conivol.R
\docType{package}
\name{conivol}
\alias{conivol}
\alias{conivol-package}
\title{conivol: A package for the (bivariate) chi-bar-squared distribution and conic intrinsic volumes}
\description{
The conivol package provides functions for the chi-bar-squared distribution,
the bivariate chi-bar-squared distribution, and the conic intrinsic volumes.
Its main function is an estimator for the weights (conic intrinsic volumes)
of the bivariate chi-bar-squared distribution from sample data,
based on the expectation maximization (EM) method.
}
\section{(Bivariate) Chi-bar-squared distribution}{

\itemize{
  \item \code{\link[conivol]{dchibarsq}} / \code{\link[conivol]{pchibarsq}} /
        \code{\link[conivol]{rchibarsq}}: evaluates the density /
        evaluates the cumulative distribution function / produces samples
        of the chi-bar-squared distribution
  \item \code{\link[conivol]{dbichibarsq}} / \code{\link[conivol]{pbichibarsq}} /
        \code{\link[conivol]{rbichibarsq}}: evaluates the density /
        evaluates the cumulative distribution function / produces samples
        of the bivariate chi-bar-squared distribution
}
}

\section{Special classes of cones}{

\itemize{
  \item \code{\link[conivol]{comp_ivols_product}}: computes the intrinsic volumes of a product cone
                                   by convolving the intrinsic volumes of
                                   its elements
  \item \code{\link[conivol]{circ_ivols}}: computes the intrinsic volumes of (a product of)
                          circular cones
  \item \code{\link[conivol]{ellips_semiax}} / \code{\link[conivol]{ellips_rbichibarsq}}:
                          computes the semiaxes / produces samples from the bivariate
                          chi-bar-squared distribution of an ellipsoidal cone
  \item \code{\link[conivol]{weyl_matrix}} / \code{\link[conivol]{weyl_ivols}}:
                          computes a matrix representation / computes the
                          intrinsic volumes of (a product of) Weyl chambers
}
}

\section{General polyhedral cones}{

\itemize{
  \item \code{\link[conivol]{polyh_reduce}}: computes a reduced representation
                     of a polyhedral cone given by generators
  \item \code{\link[conivol]{polyh_rivols_gen}} / \code{\link[conivol]{polyh_rivols_ineq}}:
                     produces samples from the intrinsic volumes distribution of
                     a polyhedral cone given by generators / inequalities
  \item \code{\link[conivol]{polyh_rbichibarsq_gen}} / \code{\link[conivol]{polyh_rbichibarsq_ineq}}:
                     produces samples from the bivariate chi-bar-squared distribution
                     with weights given by the conic intrinsic volumes of
                     a polyhedral cone given by generators / inequalities
  \item \code{\link[conivol]{polyh_bayes}}: generates functions for computing quantiles of marginals
                     of the posterior distribution and for sampling
                     from the posterior distribution,
                     given samples of the intrinsic volumes distribution
                     (based on analytic solution)
  \item \code{\link[conivol]{polyh_stan}}: generates inputs for Stan
                     (data list and model string or external file)
                     for sampling from the posterior distribution,
                     given samples of the intrinsic volumes distribution
                     using a model that naturally implies log-concavity
                     (and cannot be solved analytically)
}
}

\section{Estimating the weights of the bivariate chi-bar-squared distribution}{

\itemize{
  \item \code{\link[conivol]{estim_statdim_var}}: estimates the statistical dimension and
                     the variance of the intrinsic volumes from samples of the
                     corresponding bivariate chi-bar-squared distribution
  \item \code{\link[conivol]{init_v}}: find an initial estimate of the weights, potentially
                     based on first and/or second moment
  \item \code{\link[conivol]{comp_loglike}}: compute the log-likelihood of a weight vector
                     for specific sample data
  \item \code{\link[conivol]{prepare_data_em}}: evaluates the sample data of the bivariate chi-bar-squared
                     data (find the corresponding chi-squared density values)
  \item \code{\link[conivol]{estim_em}}: produces EM-type iterates that may or may not converge
                     to the maximum likelihood estimate for the weights
                     of the bivariate chi-bar-squared distribution
                     from sample data
  \item \code{\link[conivol]{estim_jags}} / \code{\link[conivol]{estim_stan}}:
                     generates inputs for JAGS / Stan (data list and model
                     string or external file) for sampling from the posterior distribution
                     of the intrinsic volumes,
                     given samples of the bivariate chi-bar-squared distribution
}
}

\section{Vignettes}{

\itemize{
  \item \href{../doc/conic-intrinsic-volumes.html}{conic-intrinsic-volumes}:
        introduces conic intrinsic volumes and (bivariate) chi-bar-squared distributions
  \item \href{../doc/estim-conic-intrinsic-volumes-with-EM.html}{estim-conic-intrinsic-volumes-with-EM}:
        describes the details of the algorithm for finding the intrinsic volumes of closed
        convex cones from samples of the associated bivariate chi-bar-squared distribution
  \item \href{../doc/bayesian.html}{bayesian}:
        describes the Bayesian approach for evaluating sampling data, either from
        the intrinsic volumes distribution or from the chi-bar-squared distribution,
        and with or without enforcing log-concavity for the intrinsic volumes
}
}

