% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_with_ivols.R
\name{estimate_statdim_var}
\alias{estimate_statdim_var}
\title{Estimate the statistical dimension and variance from sample data}
\usage{
estimate_statdim_var(d, m_samp)
}
\arguments{
\item{d}{the dimension of the bivariate chi-bar squared distribution.}

\item{m_samp}{two-column matrix whose rows from iid samples from the bivariate
chi-bar-squared distribution corresponding to the cone}
}
\value{
The output of \code{estimate_statdim_var} is a two-element list
        consisting of the esimated statistical dimension \code{delta}
        and variance \code{var},
}
\description{
\code{estimate_statdim_var} estimates the statistical dimension and the
variance of intrinsic volumes from samples from the corresponding
bivariate chi-bar-squared distribution.
}
\section{See also}{

\code{\link[conivol]{rbichibarsq}}, \code{\link[conivol]{circ_rbichibarsq}},
\code{\link[conivol]{polyh_rbichibarsq}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
m_samp <- circ_rbichibarsq(10^4,10,pi/8)
estimate_statdim_var(10, m_samp)

}
