% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weights.R
\name{init_v}
\alias{init_v}
\title{Finding an initial estimate of the intrinsic volumes.}
\usage{
init_v(d, init_mode = 0, delta = d/2, var = d/4)
}
\arguments{
\item{d}{the dimension of the bivariate chi-bar squared distribution.}

\item{init_mode}{specifies the way through which the initial estimate is found:
\describe{
  \item{\code{init_mode==0}:}{uniform distribution}
  \item{\code{init_mode==1}:}{discretized normal distribution}
  \item{\code{init_mode==2}:}{circular cone fitting statdim}
  \item{\code{init_mode==3}:}{circular cone fitting variance}
  \item{\code{init_mode==4}:}{circular cone fitting statdim and variance (geometric mean)}
}}

\item{delta}{an estimate of the statistical dimension of the cone.}

\item{var}{an estimate of the variane of the intrinsic volumes.}
}
\value{
The output of \code{init_v} is a \code{(d+1)}-column vector.
}
\description{
\code{init_v} find an initial estimate of the intrinsic volumes via
moment-fitting.
}
\section{See also}{

\code{\link[conivol]{rbichibarsq}}, \code{\link[conivol]{rbichibarsq_circ}},
\code{\link[conivol]{rbichibarsq_polyh}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
m_samp <- rbichibarsq_circ(10^6,c(5,5),c(pi/3,pi/4))
est <- estimate_statdim_var(d, m_samp)
init_v( 10 )
init_v( 10, 1, delta=est$delta, var=est$var )
init_v( 10, 2, delta=est$delta )
init_v( 10, 3, delta=est$delta, var=est$var )
init_v( 10, 4, delta=est$delta, var=est$var )

}
