% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_weights.R
\name{prepare_data}
\alias{prepare_data}
\title{Evaluate the sample data for maximum likelihood estimation.}
\usage{
prepare_data(d, m_samp)
}
\arguments{
\item{d}{the dimension of the bivariate chi-bar squared distribution.}

\item{m_samp}{two-column matrix whose rows from iid samples from a bivariate
chi-bar-squared distribution.}
}
\value{
The output of \code{prepare_data} is a list of four elements:
        \describe{
          \item{\code{n}:}{the number of overall sample points
                           (including those in primal or polar cone)}
          \item{\code{prop_prim}:}{proportion of points in primal cone}
          \item{\code{prop_pol}:}{proportion of points}
          \item{\code{dens}:}{the density values of the effective sample points
                              (neither in primal nor polar cone); that is,
                              \code{dens} is a \code{(d-1)} row matrix
                              such that the \code{k}th row contains the products
                              of the density values of the chi_k^2 and chi_(d-k)^2
                              distributions evaluated in the effective sample points;
                              the row-form of the matrix is more convenient for
                              the computations}
        }
}
\description{
\code{prepare_data} takes a two-column matrix whose rows form
iid samples from a bivariate chi-bar-squared distribution and
prepares the data used in maximum likelihood estimation.
}
\section{See also}{

\code{\link[conivol]{rbichibarsq}}, \code{\link[conivol]{circ_rbichibarsq}},
\code{\link[conivol]{rbichibarsq_polyh}}, \code{\link[conivol]{find_ivols_EM}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
m_samp <- circ_rbichibarsq(10^6,c(5,5),c(pi/3,pi/4))
prepare_data( 10, m_samp )

}
