% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/circular_cones.R
\name{rbichibarsq_circ}
\alias{rbichibarsq_circ}
\title{Sample from bivariate chi-bar-squared distribution of products of circular cones}
\usage{
rbichibarsq_circ(n, d, alpha)
}
\arguments{
\item{n}{number of samples}

\item{d}{dimension}

\item{alpha}{angle}
}
\value{
The output of \code{rbichibarsq_circ(n,d,alpha)} is a
        \code{n} by \code{2} matrix such that the rows form iid samples from the
        bivariate chi-bar-squared distribution with weights given by the
        intrinsic volumes of a product of circular cones corresponding
        to the vector of dimensions \code{d} and the vector of angles \code{alpha}.
}
\description{
\code{rbichibarsq_circ} generates an \code{n} by \code{2} matrix
such that the rows form iid samples from the bivariate chi-bar-squared
distribution of products of circular cones.
}
\section{See also}{

\code{\link[conivol]{circ_ivol}}, \code{\link[conivol]{rbichibarsq_polyh}},
\code{\link[conivol]{rbichibarsq}}, \code{\link[conivol]{rchibarsq}},
\code{\link[stats]{rchisq}},

Package: \code{\link[conivol]{conivol}}
}

\examples{
rbichibarsq_circ(20,5,pi/3)
rbichibarsq_circ(20,c(5,3),c(pi/3,pi/4))

}
