% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedral.R
\name{rbichibarsq_polyh_ineq}
\alias{rbichibarsq_polyh_ineq}
\title{Sample from bivariate chi-bar-squared distribution of a polyhedral cone given by inequalities}
\usage{
rbichibarsq_polyh_ineq(n, A, reduce = TRUE, shrink = TRUE)
}
\arguments{
\item{n}{number of samples}

\item{A}{matrix}

\item{reduce}{logical; if \code{TRUE}, the cone defined by \code{A} will be
decomposed orthogonally w.r.t. its lineality space}
}
\value{
The output of \code{rbichibarsq_polyh_ineq(n,A)}, with the default value
        \code{reduce==TRUE}, is a list containing the following elements:
\itemize{
  \item \code{dim}: the dimension of the linear span of \code{C},
  \item \code{lin}: the lineality of the cone \code{C},
  \item \code{QL}: an orthogonal basis of the orthogonal complement of the
                   linear span of \code{C},
                   set to \code{NA} if \code{dim(C)==d},
  \item \code{QC}: an orthogonal basis of the projection of \code{C} onto
                   the orthogonal complement of the lineality space of \code{C},
                   set to \code{NA} if \code{C} is a linear space,
  \item \code{A_reduced}: a matrix defining the reduced cone,
  \item \code{samples}: an \code{n} by \code{2} matrix rows of the matrix form
        iid samples from the bivariate chi-bar-squared distribution with
        weights given by the intrinsic volumes of the reduced cone \code{{y|A_reduced^T y<=0}}.
}
        If \code{reduce==FALSE} then the output is only an
        \code{n} by \code{2} matrix such that its rows form
        iid samples from the bivariate chi-bar-squared distribution with
        weights given by the intrinsic volumes of the cone \code{{y|A^Ty<=0}},
        set to \code{NA} if \code{C} is a linear space.
}
\description{
\code{rbichibarsq_polyh_ineq} generates an \code{n} by \code{2} matrix
such that the rows form iid samples from the bivariate chi-bar-squared
distribution of the polyhedral cone given by inequalities, that is, in the
form \code{{y|A^Ty<=0}}. If \code{reduce==TRUE}, which is the default, then a
reduced form of the cone will be computed and the bivariate chi-bar-squared
distribution will correspond to the reduced form, and the output will contain
further elements (in form of a list), see below.
}
\note{
See \href{../doc/conic-intrinsic-volumes.html#sampling_polyh}{this vignette}
      for further info.
}
\section{See also}{

\code{\link[conivol]{rbichibarsq_polyh_gen}}, \code{\link[conivol]{rbichibarsq}},
\code{\link[conivol]{rbichibarsq_circ}}

Package: \code{\link[conivol]{conivol}}
}

\examples{
rbichibarsq_polyh_ineq(20,matrix(1:12,4,3)

}
