% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyhedral.R
\name{remove_redundancies}
\alias{remove_redundancies}
\title{Remove the columns of a matrix that are not extreme rays of the cone generated by the matrix}
\usage{
remove_redundancies(A, tol = 1e-08)
}
\arguments{
\item{A}{matrix}

\item{tol}{tolerance (single precision machine epsilon by default)}
}
\value{
The output consists of a reduced matrix \code{A_red} with 
        \code{{Ax|x>=0} = {A_red x|x>=0}}, and so that no proper submatrix of \code{A_red}
        satisfies this property.
}
\description{
\code{remove_redundancies} takes as input a \code{n} by \code{m} matrix \code{A} and returns a 
submatrix \code{A_red} consisting of a minimal subset of columns of \code{A} that generate
the convex cone \code{{Ax|x>=0}}.
}
\note{
The function uses non-negative least squres (nnls) to determine, for each column of A,
      whether it is contained in the conic hull of the remaining columns.
}
\examples{
A <- cbind(diag(1,3),diag(1,3)+matrix(1,ncol=3,nrow=3))
A <- A[,sample(ncol(A))]
print(A)
A_red <- remove_redundancies(A)
print(A_red)

}
